requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
python-decouple==3.8
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
psycopg2==2.9.10; sys_platform == 'windows'
psycopg2-binary==2.9.10; sys_platform == 'linux' or sys_platform == "darwin"
PyJWT==2.10.1
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
django-cors-headers==4.6.0
django-dotenv==1.4.2
django-guardian==2.4.0